Add files via upload
In this data engineering project, a dataset related to the gaming industry is utilized. The dataset is stored in an AWS S3 bucket and is mounted to a Databricks workspace. Using Databricks, a Spark DataFrame is generated from the dataset, and SparkSQL is used to analyze the data. Various queries are performed on the DataFrame to extract insights and patterns from the data.